move bytes from streamable to chia_rs root
