+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod bytes;
 pub mod error;
 pub mod gen;
 pub mod merkle_set;
