+++ src/streamable/mod.rs
@@ -1,4 +1,3 @@
-pub mod bytes;
 // pub mod fullblock;
 // pub mod program;
 pub mod de;
